Upload improved realistic car-like vehicle mod #15
The new uploaded car is a more real-life replication of a sedan, and is now able to follow the physics of the SMCE default car, and is able to correctly move and turn in the terrain, with sensors working and intact. The sensor's locations also had to be slightly adjusted, as the new vehicle mod has different measurements.