Fix issue of env not registered + trying more sensors #15
From the previous commit, the .gd file that is responsible for
initialising the vehicle and the terrain did not work anymore, so it had
to be stored in a specific directory called "mods" for it to function
again normally.

Some experimentation was done in how more sensors can be added to the
car when needed, so 2 extra slots called "FrontLeft" and "FrontRight"
were created, and each have an IR sensor located there, for the time
being.