Configure movement methods in the android app
Configred the movement methods in order to better reflect the MQTT Topic
structure. Also created a new section in the smartcar.ino file for the
MQTT methods, along with creating constants for speed limits and
steering angles. Overall, this lays a good foundation for better car
control over the MQTT Protocol, as well as laying the ground work for a
a future joystick implementation.