Initialise OV767X cam in Arduino sketch as prelim
The OV767X library was used, and the Camera object was initialised in
the void setup() method, and loop() also. Camera is currently located at the
front hood of the car.

NEXT STEP: Broadcast the captured footage from the cam in the Android
app, which was also updated from master for latest version to be done
work on.

Co-authored-by: Gabri1998 <gusalgaah@student.gu.se>