add the number picker and connect with the toggel button #68
toggle button wasn't suitable because we had to click on toggle button (to ON the cruise control)and same time we had to move stick_up,that wasn't wrokinh on the emulato. We add the number picker, so now  we can select the spcific speed and then we can ON the cruise control button.