Set up birdeye cam #64
The camera is now added in the mod, adjusted to a suitable height so
that the entire car and its surrounding corners are visible throughout.
Vehicle_config changes are done for declaring the camera to the
placeholder in the mod called "360Camera". It is now to be tested in the
app, with details to follow in the pull request 105.